+++ src/Repository/Repository.php
@@ -2,10 +2,14 @@
 
 namespace App\Repository;
 
+use App\Service\HTMLSanitizerService;
+use App\Service\ServerInformationService;
 use Cake\Database\Connection;
 use Cake\Database\Query;
 use DateTime;
+use ParagonIE\EasyDB\EasyDB;
 use RuntimeException;
+use DI\Attribute\Inject;
 
 /**
  * Factory.
@@ -14,6 +18,12 @@ class Repository
 {
     protected Connection $connection;
 
+    protected $db = null;
+
+    public ?string $entityClass = null;
+
+    private $purifier;
+
     public array $timestampedColumns = [
         'expiration',
         'bantime',
@@ -21,17 +31,32 @@ class Repository
         'initialize_datetime',
         'start_datetime',
         'shutdown_datetime',
-        'end_datetime'
+        'end_datetime',
+        'timestamp'
     ];
 
-    /**
-     * The constructor.
-     *
-     * @param Connection $connection The database connection
-     */
-    public function __construct(Connection $connection)
+    public array $serverPortColumns = [
+        'port',
+        'server_port'
+    ];
+
+    public array $stripHTMLColumns = [
+        'message'
+    ];
+
+    public $results = null;
+    public ?int $pages = null;
+
+    private ?array $servers = null;
+
+    #[Inject]
+    private HTMLSanitizerService $html;
+
+    public function __construct(Connection $connection, EasyDB $db)
     {
         $this->connection = $connection;
+        $this->db = $db;
+
     }
 
     /**
@@ -98,9 +123,8 @@ class Repository
         return $this->newQuery()->delete($table);
     }
 
-    protected function parseTimestamps(array $data): array
+    protected function parseTimestamps(object|array $data): object|array
     {
-
         foreach($data as $k => &$v) {
             if(in_array($k, $this->timestampedColumns)) {
                 if(is_null($v)) {
@@ -111,5 +135,68 @@ class Repository
             }
         }
         return $data;
+    }
+
+    protected function mapServer(object|array $data): object|array
+    {
+        if(!$this->servers) {
+            $this->servers = ServerInformationService::getServerInfo();
+        }
+        foreach($data as $k => &$v) {
+            if(in_array($k, $this->serverPortColumns)) {
+                if(is_null($v)) {
+                    $v = null;
+                } else {
+                    $data->server = ServerInformationService::getServerFromPort($v, $this->servers);
+                }
+            }
+        }
+        return $data;
+    }
+
+    protected function stripHTML(object|array $data): object|array
+    {
+        foreach($data as $k => &$v) {
+            if(in_array($k, $this->stripHTMLColumns)) {
+                if(is_null($v)) {
+                    $v = null;
+                } else {
+                    $v = $this->html->sanitizeString($v);
+                }
+            }
+        }
+        return $data;
+    }
+
+    protected function setResults($results): self
+    {
+        if(is_array($results)) {
+            foreach($results as &$r) {
+                //These should be attributes
+                $r = $this->parseTimestamps($r);
+                $r = $this->mapServer($r);
+                $r = $this->stripHTML($r);
+                if($this->entityClass && class_exists($this->entityClass)) {
+                    $r = new $this->entityClass(...array_values((array) $r));
+                }
+            }
+        }
+
+        $this->results = $results;
+        return $this;
+    }
+
+    public function getResults(): array|object|null
+    {
+        return $this->results;
+    }
+    public function setPages(int $pages): self
+    {
+        $this->pages = $pages;
+        return $this;
+    }
+    public function getPages(): int
+    {
+        return $this->pages;
     }
 }